LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Close-Reason-124Bytes.any-expected.txt

CONSOLE MESSAGE: WebSocket close message is too long.

PASS Create WebSocket - Close the Connection - close(code, 'reason more than 123 bytes') - SYNTAX_ERR is thrown
